Fixed security issue
We're using the net.liftweb.util.Helpers.nextFuncName function to
generate unique names for the input fields. This has the advantage
of beeing more safe against potential hackers. However, the previous
implementation generated these names upon construction of the Formlets -
this is no good if you have your forms defined in an object as the names
would be the same during the entire life-span of the application.

To fix this formlets are now () => (NodeSeq, Env => Either[Error, A], Names)
so the names are generated each time to form is displayed